# docs/architecture/api_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Daily Stock Analysis API",
    "description": "KR/US Watchlist AI Stock Analysis API\n\n## Modules\n- Stock analysis: trigger AI analysis\n- History: query historical analysis reports\n- Stock data: fetch market quotes\n\n## Authentication\nNo authentication required in the current version",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/history": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "히스토리 분석 목록 조회",
        "description": "히스토리 분석 기록 요약을 페이지 단위로 조회하며, 종목 코드와 날짜 범위로 필터링할 수 있습니다.",
        "operationId": "get_history_list_api_v1_history_get",
        "parameters": [
          {
            "name": "stock_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "종목 코드 필터",
              "title": "Stock Code"
            },
            "description": "종목 코드 필터"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "시작일 (YYYY-MM-DD)",
              "title": "Start Date"
            },
            "description": "시작일 (YYYY-MM-DD)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "종료일 (YYYY-MM-DD)",
              "title": "End Date"
            },
            "description": "종료일 (YYYY-MM-DD)"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "페이지 번호 (1부터 시작)",
              "default": 1,
              "title": "Page"
            },
            "description": "페이지 번호 (1부터 시작)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "페이지당 항목 수",
              "default": 20,
              "title": "Limit"
            },
            "description": "페이지당 항목 수"
          }
        ],
        "responses": {
          "200": {
            "description": "히스토리 기록 목록",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryListResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/history/{record_id}": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "히스토리 보고서 상세 조회",
        "description": "분석 히스토리 기록 ID 또는 query_id로 전체 히스토리 분석 보고서를 조회합니다.",
        "operationId": "get_history_detail_api_v1_history__record_id__get",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "보고서 상세",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisReport"
                }
              }
            }
          },
          "404": {
            "description": "보고서를 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/history/{record_id}/news": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "히스토리 보고서 관련 뉴스 조회",
        "description": "분석 히스토리 기록 ID로 관련 뉴스 인텔리전스 목록을 조회합니다. 결과가 없어도 200을 반환합니다.",
        "operationId": "get_history_news_api_v1_history__record_id__news_get",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "반환 항목 수 제한",
              "default": 20,
              "title": "Limit"
            },
            "description": "반환 항목 수 제한"
          }
        ],
        "responses": {
          "200": {
            "description": "뉴스 인텔리전스 목록",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsIntelResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/stocks/{stock_code}/history": {
      "get": {
        "tags": [
          "Stocks"
        ],
        "summary": "종목 과거 시세 조회",
        "description": "지정한 종목의 과거 OHLCV 데이터를 조회합니다",
        "operationId": "get_stock_history_api_v1_stocks__stock_code__history_get",
        "parameters": [
          {
            "name": "stock_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Stock Code"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(daily|weekly|monthly)$",
              "description": "가격 주기",
              "default": "daily",
              "title": "Period"
            },
            "description": "가격 주기"
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "description": "조회 일수",
              "default": 30,
              "title": "Days"
            },
            "description": "조회 일수"
          }
        ],
        "responses": {
          "200": {
            "description": "과거 시세 데이터",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "지원하지 않는 기간 파라미터",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisReport": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/ReportMeta",
            "description": "메타 정보"
          },
          "summary": {
            "$ref": "#/components/schemas/ReportSummary",
            "description": "요약 영역"
          },
          "strategy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportStrategy"
              },
              {
                "type": "null"
              }
            ],
            "description": "전략 가격 영역"
          },
          "details": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportDetails"
              },
              {
                "type": "null"
              }
            ],
            "description": "상세 영역"
          }
        },
        "type": "object",
        "required": [
          "meta",
          "summary"
        ],
        "title": "AnalysisReport",
        "description": "Complete analysis report.",
        "example": {
          "meta": {
            "created_at": "2024-01-01T12:00:00",
            "query_id": "abc123",
            "report_type": "detailed",
            "stock_code": "005930",
            "stock_name": "삼성전자"
          },
          "strategy": {
            "ideal_buy": "1800.00",
            "secondary_buy": "1750.00",
            "stop_loss": "1700.00",
            "take_profit": "2000.00"
          },
          "summary": {
            "analysis_summary": "기술적 흐름이 양호해 보유 관점이 유효합니다",
            "operation_advice": "보유",
            "sentiment_label": "낙관",
            "sentiment_score": 75,
            "trend_prediction": "강세"
          }
        }
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Error type",
            "example": "validation_error"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Error detail",
            "example": "요청 파라미터 오류"
          },
          "detail": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "Additional error information"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse",
        "description": "Error response.",
        "example": {
          "error": "not_found",
          "message": "리소스를 찾을 수 없습니다"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HistoryItem": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id",
            "description": "분석 이력 기본 키 ID"
          },
          "query_id": {
            "type": "string",
            "title": "Query Id",
            "description": "분석 기록의 query_id"
          },
          "stock_code": {
            "type": "string",
            "title": "Stock Code",
            "description": "종목 코드"
          },
          "stock_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stock Name",
            "description": "종목명"
          },
          "report_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Type",
            "description": "리포트 유형"
          },
          "sentiment_score": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentiment Score",
            "description": "심리 점수 (0-100)"
          },
          "operation_advice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Advice",
            "description": "투자 판단"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At",
            "description": "생성 시간"
          }
        },
        "type": "object",
        "required": [
          "query_id",
          "stock_code"
        ],
        "title": "HistoryItem",
        "description": "History summary item used in list views.",
        "example": {
          "created_at": "2024-01-01T12:00:00",
          "id": 1234,
          "operation_advice": "보유",
          "query_id": "abc123",
          "report_type": "detailed",
          "sentiment_score": 75,
          "stock_code": "005930",
          "stock_name": "삼성전자"
        }
      },
      "HistoryListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "전체 기록 수"
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "description": "현재 페이지"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "페이지당 개수"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/HistoryItem"
            },
            "type": "array",
            "title": "Items",
            "description": "기록 목록"
          }
        },
        "type": "object",
        "required": [
          "total",
          "page",
          "limit"
        ],
        "title": "HistoryListResponse",
        "description": "History list response.",
        "example": {
          "items": [],
          "limit": 20,
          "page": 1,
          "total": 100
        }
      },
...
      "NewsIntelItem": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "description": "뉴스 제목"
          },
          "snippet": {
            "type": "string",
            "title": "Snippet",
            "description": "뉴스 요약",
            "default": ""
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "뉴스 링크"
          }
        },
        "type": "object",
        "required": [
          "title",
          "url"
        ],
        "title": "NewsIntelItem",
        "description": "News intelligence item.",
        "example": {
          "snippet": "삼성전자가 분기 실적과 향후 전망을 공개했습니다...",
          "title": "삼성전자, 분기 실적 발표",
          "url": "https://example.com/news/123"
        }
      },
      "NewsIntelResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "뉴스 수"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/NewsIntelItem"
            },
            "type": "array",
            "title": "Items",
            "description": "뉴스 목록"
          }
        },
        "type": "object",
        "required": [
          "total"
        ],
        "title": "NewsIntelResponse",
        "description": "News intelligence response.",
        "example": {
          "items": [],
          "total": 2
        }
      },
...
      "ReportDetails": {
        "properties": {
          "news_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "News Content",
            "description": "뉴스 요약"
          },
          "raw_result": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Raw Result",
            "description": "원본 분석 결과(JSON)"
          },
          "context_snapshot": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Context Snapshot",
            "description": "분석 시점 컨텍스트 스냅샷(JSON)"
          }
        },
        "type": "object",
        "title": "ReportDetails",
        "description": "Report detail section."
      },
      "ReportMeta": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id",
            "description": "분석 이력 기본 키 ID"
          },
          "query_id": {
            "type": "string",
            "title": "Query Id",
            "description": "분석 기록의 query_id"
          },
          "stock_code": {
            "type": "string",
            "title": "Stock Code",
            "description": "종목 코드"
          },
          "stock_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stock Name",
            "description": "종목명"
          },
          "report_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Type",
            "description": "리포트 유형"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At",
            "description": "생성 시간"
          },
          "current_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Price",
            "description": "분석 시점 가격"
          },
          "change_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Pct",
            "description": "분석 시점 등락률(%)"
          },
          "model_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Used",
            "description": "분석에 사용한 LLM 모델"
          }
        },
        "type": "object",
        "required": [
          "query_id",
          "stock_code"
        ],
        "title": "ReportMeta",
        "description": "Report metadata."
      },
      "ReportStrategy": {
        "properties": {
          "ideal_buy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ideal Buy",
            "description": "우선 매수가"
          },
          "secondary_buy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secondary Buy",
            "description": "보조 매수가"
          },
          "stop_loss": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Loss",
            "description": "손절가"
          },
          "take_profit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Take Profit",
            "description": "목표가"
          }
        },
        "type": "object",
        "title": "ReportStrategy",
        "description": "Strategy price section."
      },
      "ReportSummary": {
        "properties": {
          "analysis_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Summary",
            "description": "핵심 결론"
          },
          "operation_advice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Advice",
            "description": "투자 판단"
          },
          "trend_prediction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trend Prediction",
            "description": "추세 전망"
          },
          "sentiment_score": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentiment Score",
            "description": "심리 점수 (0-100)"
          },
          "sentiment_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentiment Label",
            "description": "심리 라벨"
          }
        },
        "type": "object",
        "title": "ReportSummary",
        "description": "Report summary section."
      },
...
      "StockHistoryResponse": {
        "properties": {
          "stock_code": {
            "type": "string",
            "title": "Stock Code",
            "description": "종목 코드"
          },
          "stock_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stock Name",
            "description": "종목명"
          },
          "period": {
            "type": "string",
            "title": "Period",
            "description": "가격 주기"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/KLineData"
            },
            "type": "array",
            "title": "Data",
            "description": "가격 데이터 목록"
          }
        },
        "type": "object",
        "required": [
          "stock_code",
          "period"
        ],
        "title": "StockHistoryResponse",
        "description": "Historical stock quote response.",
        "example": {
          "data": [],
          "period": "daily",
          "stock_code": "AAPL",
          "stock_name": "Apple Inc."
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}